requirements.txt
pandas==2.2.3
numpy==2.1.3
scikit-learn==1.5.2
joblib==1.4.2
jieba==0.42.1
pyyaml==6.0.2
tqdm==4.67.1
streamlit==1.41.1